Add python_introduction file

+++ python_introduction/simple_interest.py
@@ -6,7 +6,7 @@ rate = 0.05 #declare interest rate as 5%
 time = 3 #declare time as 3 years
 
 # Calculate simple interest using the formula: P * R * T
-interest = float(principal * rate * time)
+interest = principal * rate * time
 
 # Display the calculated interest to the user
 print ("The simple interest is: ", interest)
